# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@karimastore.com
    name: Karima Store API Support
  description: |-
    Karima Store E-commerce API with JWT Authentication

    ## Authentication

    This API uses **JWT (JSON Web Token)** for authentication.

    ### For Web/Browser Clients:
    1. Login via /api/v1/auth/login endpoint
    2. Session cookie (jwt) will be set automatically
    3. Make API requests with the cookie included

    ### For API/Mobile Clients:
    1. Obtain JWT token from /api/v1/auth/login endpoint
    2. Include token in requests:
    - Method 1: Authorization: Bearer <jwt_token>
    - Method 2: X-Session-Token: <jwt_token> header

    ### Authorization Levels:
    - **Public**: No authentication required (GET endpoints for browsing)
    - **Authenticated**: Valid JWT token required
    - **Admin**: Valid JWT token + admin role
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Karima Store API
  version: "1.0"
paths:
...
  /api/v1/products:
    get:
      description: Retrieve a list of products with pagination and optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Minimum price
        in: query
        name: min_price
        type: number
      - description: Maximum price
        in: query
        name: max_price
        type: number
      - description: Filter by brand
        in: query
        name: brand
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: 'Create a new product with the provided details. **Admin only**:
        Requires authentication with admin role.'
      parameters:
      - description: Product object
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: No valid session or session expired'
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 'Forbidden: Insufficient permissions (admin role required)'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Create a new product (Admin only)
      tags:
      - products
  /api/v1/products/:product_id/media:
    get:
      consumes:
      - application/json
      description: Retrieve all media files for a specific product
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Media list
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get media by product
      tags:
      - media
...
  /api/v1/products/{id}:
    delete:
      description: 'Delete a product by ID. **Admin only**: Requires authentication
        with admin role.'
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: No valid session or session expired'
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 'Forbidden: Insufficient permissions (admin role required)'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Delete product (Admin only)
      tags:
      - products
    get:
      description: Retrieve a single product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get product by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: 'Update an existing product by ID. **Admin only**: Requires authentication
        with admin role.'
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product object
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: No valid session or session expired'
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 'Forbidden: Insufficient permissions (admin role required)'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Update product (Admin only)
      tags:
      - products
  /api/v1/products/{id}/media:
    get:
      consumes:
      - application/json
      description: Retrieve all media files for a specific product
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Media list
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get product media
      tags:
      - products
    post:
      consumes:
      - multipart/form-data
      description: 'Upload an image file for a product. **Admin only**: Requires authentication
        with admin role.'
      parameters:
      - description: Product ID
        in: formData
        name: product_id
        required: true
        type: integer
      - description: Image file
        in: formData
        name: file
        required: true
        type: file
      - description: Position
        in: formData
        name: position
        type: integer
      - description: Is primary
        in: formData
        name: is_primary
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Upload response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: No valid session or session expired'
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 'Forbidden: Insufficient permissions (admin role required)'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Upload product media (Admin only)
      tags:
      - products
  /api/v1/products/{id}/stock:
    patch:
      consumes:
      - application/json
      description: 'Update the stock quantity for a product. **Admin only**: Requires
        authentication with admin role.'
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Stock update object
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: No valid session or session expired'
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 'Forbidden: Insufficient permissions (admin role required)'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Update product stock (Admin only)
      tags:
      - products
...
  /api/v1/products/bestsellers:
    get:
      description: Retrieve best-selling products based on sales count
      parameters:
      - default: 10
        description: Number of products to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get best sellers
      tags:
      - products
  /api/v1/products/category/{category}:
    get:
      description: Retrieve products filtered by category
      parameters:
      - description: Product category
        in: path
        name: category
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get products by category
      tags:
      - products
  /api/v1/products/featured:
    get:
      description: Retrieve featured products based on views and sales
      parameters:
      - default: 10
        description: Number of products to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get featured products
      tags:
      - products
  /api/v1/products/search:
    get:
      description: Search for products by name, description, or brand
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Search products
      tags:
      - products
  /api/v1/products/slug/{slug}:
    get:
      description: Retrieve a single product by its slug
      parameters:
      - description: Product slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get product by slug
      tags:
      - products
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"